+++ source/include/microapp/cs_MicroappProtocol.h
@@ -2,7 +2,6 @@
 
 #include <cs_MicroappStructs.h>
 #include <events/cs_EventListener.h>
-#include <structs/buffer/cs_CircularBuffer.h>
 
 extern "C" {
 #include <util/cs_DoubleStackCoroutine.h>
@@ -37,12 +36,6 @@ struct microapp_ble_isr_t {
 	bool registered;
 };
 
-struct __attribute__((packed)) microapp_buffered_mesh_message_t {
-	stone_id_t stoneId;  // Stone ID of the sender.
-	uint8_t messageSize;
-	uint8_t message[MICROAPP_MAX_MESH_MESSAGE_SIZE];
-};
-
 /**
  * The class MicroappProtocol has functionality to store a second app (and perhaps in the future even more apps) on
  * another part of the flash memory.
@@ -65,11 +58,6 @@ private:
 	 * Limit maximum number of to be registered service routines for the microapp.
 	 */
 	static const uint8_t MICROAPP_MAX_BLE_ISR_COUNT = 8;
-
-	/**
-	 * Max number of mesh messages that will be queued.
-	 */
-	const uint8_t MICROAPP_MAX_MESH_MESSAGES_BUFFERED = 3;
 
 	/**
 	 * Limit the number of callbacks in a tick (if -1) there is no limit.
@@ -108,36 +96,24 @@ private:
 	uint8_t _callCounter;
 
 	/**
-	 * Store data for callbacks towards the microapp.
-	 */
-	CircularBuffer<uint8_t>* _callbackData;
-
-	/**
-	 * Flag to indicate whether to forward scanned devices to microapp
-	 */
-	bool _microappIsScanning;
-
-	/**
-	 * Buffer received mesh messages.
-	 *
-	 * Starts with microapp_buffered_mesh_message_header_t, followed by the message.
-	 */
-	CircularBuffer<microapp_buffered_mesh_message_t> _meshMessageBuffer;
-
-	/**
 	 * Soft interrupt counter
 	 */
 	int8_t _softInterruptCounter;
 
 	/**
-	 * Number of empty interrupt slots
-	 */
-	int8_t _emptyInterruptSlots;
-
-	/**
 	 * To throttle the ticks themselves.
 	 */
 	uint8_t _tickCounter;
+	
+	/**
+	 * Maps digital pins to interrupts.
+	 */
+	int digitalPinToInterrupt(int pin);
+	
+	/**
+	 * Flag to indicate whether to forward scanned devices to microapp.
+	 */
+	bool _microappIsScanning;
 
 protected:
 	/**
@@ -181,11 +157,6 @@ protected:
 	void softInterruptBle(scanned_device_t* dev);
 
 	/**
-	 * A soft interrupt is already in progress.
-	 */
-	bool softInterruptInProgress();
-
-	/**
 	 * Get incoming microapp buffer (from coargs).
 	 */
 	uint8_t* getInputMicroappBuffer();
@@ -194,11 +165,6 @@ protected:
 	 * Get outgoing microapp buffer (from coargs).
 	 */
 	uint8_t* getOutputMicroappBuffer();
-
-	/**
-	 * Handle a received mesh message.
-	 */
-	void onMeshMessage(MeshMsgEvent event);
 
 	/**
 	 * Handle a scanned BLE device.
@@ -212,70 +178,6 @@ protected:
 	 * @param[in] cmd                            Any microapp command
 	 */
 	bool stopAfterMicroappCommand(microapp_cmd_t* cmd);
-
-	/**
-	 * Handle microapp log commands.
-	 *
-	 * @param[in] cmd                            A microapp command for logging to serial.
-	 */
-	cs_ret_code_t handleMicroappLogCommand(microapp_log_cmd_t* cmd);
-
-	/**
-	 * Handle microapp delay commands.
-	 *
-	 * @param[in] cmd                            A microapp command to introduce a delay for the microapp itself.
-	 */
-	cs_ret_code_t handleMicroappDelayCommand(microapp_delay_cmd_t* cmd);
-
-	/**
-	 * Handle microapp pin commands.
-	 */
-	cs_ret_code_t handleMicroappPinCommand(microapp_pin_cmd_t* cmd);
-
-	/**
-	 * Handle microapp pin switching commands.
-	 */
-	cs_ret_code_t handleMicroappPinSwitchCommand(microapp_pin_cmd_t* cmd);
-
-	/**
-	 * Handle microapp pin commands for setting pin modes.
-	 */
-	cs_ret_code_t handleMicroappPinSetModeCommand(microapp_pin_cmd_t* cmd);
-
-	/**
-	 * Handle microapp pin commands for reading and writing pins.
-	 */
-	cs_ret_code_t handleMicroappPinActionCommand(microapp_pin_cmd_t* cmd);
-
-	/**
-	 * Handle microapp commands for advertising microapp state in service data.
-	 */
-	cs_ret_code_t handleMicroappServiceDataCommand(microapp_service_data_cmd_t* cmd);
-
-	/**
-	 * Handle microapp TWI commands.
-	 */
-	cs_ret_code_t handleMicroappTwiCommand(microapp_twi_cmd_t* cmd);
-
-	/**
-	 * Handle microapp BLE commands.
-	 */
-	cs_ret_code_t handleMicroappBleCommand(microapp_ble_cmd_t* cmd);
-
-	/**
-	 * Handle microapp commands for power usage requests.
-	 */
-	cs_ret_code_t handleMicroappPowerUsageCommand(microapp_power_usage_cmd_t* cmd);
-
-	/**
-	 * Handle microapp commands for presence requests.
-	 */
-	cs_ret_code_t handleMicroappPresenceCommand(microapp_presence_cmd_t* cmd);
-
-	/**
-	 * Handle microapp commands for sending and reading mesh messages.
-	 */
-	cs_ret_code_t handleMicroappMeshCommand(microapp_mesh_cmd_t* cmd);
 
 public:
 	static MicroappProtocol& getInstance() {
@@ -299,12 +201,13 @@ public:
 	void tickMicroapp(uint8_t appIndex);
 
 	/**
+	 * Shortcut setScanning for MicroappCommandHandler to update state of this instance.
+	 */
+	void setScanning(bool scanning);
+
+	/**
 	 * Receive events (for example for i2c)
 	 */
 	void handleEvent(event_t& event);
 
-	/**
-	 * Handle commands from the microapp
-	 */
-	cs_ret_code_t handleMicroappCommand(microapp_cmd_t* cmd);
 };
